COM-00094, DEV-11113, EIA3216 Cap
+ EIA3216 Capacitor to category "Created Models" in "Bare Components"
+ Crystal SMD 16MHz COM-00094 to category "Components"
+ Arduino Pro Mini 328 - 5V:16MHz DEV-11113 to category "Dev Tools"

Made the capacitor for the Pro Mini, as well as the Oscillator, but
found out the Pro Mini doesn't use that OSC, so put in a rect for the
Resonator.  The parts on the Pro Mini were put on in the locations that
were specified in the EAGLE Files.  It should be acurate to a 100th of
a millimeter.